first crack at modular code
